+++ app/(consumer)/talents/[slug]/_components/IconCard.tsx
