+++ app/(consumer)/talents/[slug]/_components/IntroduceCard.tsx
