+++ app/(consumer)/talents/[slug]/_components/PorfolioCard.tsx
