+++ app/(consumer)/talents/[slug]/_components/ResumeCard.tsx
