+++ app/(consumer)/talents/register/_components/CareerComponent.tsx
