+++ app/(consumer)/talents/register/_components/CodeRegisterComponent.tsx
