+++ app/(consumer)/talents/register/_components/EducationComponent.tsx
