+++ app/(consumer)/talents/register/_components/IntroComponent.tsx
