+++ app/(consumer)/talents/register/_components/LinkComponent.tsx
