+++ app/(consumer)/talents/register/_components/PhotoComponent.tsx
