+++ app/(consumer)/talents/register/_components/QualificationComponent.tsx
