+++ app/(consumer)/talents/register/_components/SkillComponent.tsx
