+++ app/(consumer)/talents/register/_components/TendencyComponent.tsx
